.Refactor move mocks into test/mocks folder
