+++ nomadWorker/upscale_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	mock_aws "github.com/thomasobenaus/sokar/test/aws"
+	mock_aws "github.com/thomasobenaus/sokar/test/mocks/aws"
 )
 
 func TestUpscale(t *testing.T) {
