+++ scaleAlertAggregator/evaluate_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"github.com/thomasobenaus/sokar/test/metrics"
+	"github.com/thomasobenaus/sokar/test/mocks/metrics"
 )
 
 func Test_IsScalingNeeded(t *testing.T) {
