+++ test/mocks/sokar/mock_scaleEventEmitter_IF.go
